Allow Placement of New Track
Player can now build track on the board!
Select a tile on the board and select an appropriate track tile to place
(including the need to rotate it as desired). Restrictions include
checking that track on the board does not get removed, track does not
point off the board or illegal placement, tile level must increase
incrementally, and some board tiles have restricted tile options.